remove class, interface - asnyc merge with sync class
