+++ src/SimpleWait.Core/IWait.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 
 namespace SimpleWait.Core
 {
@@ -35,5 +36,15 @@ namespace SimpleWait.Core
         /// If TResult is an object, the method returns the object when the condition evaluates to a value other than <see langword="null"/>.</returns>
         /// <exception cref="ArgumentException">Thrown when TResult is not boolean or an object type.</exception>
         TResult Until<TResult>(Func<T, TResult> condition);
+
+        /// <summary>
+        /// Waits until a condition is true or times out.
+        /// </summary>
+        /// <typeparam name="TResult">The type of result to expect from the condition.</typeparam>
+        /// <param name="condition">A delegate taking a TSource as its parameter, and returning a TResult.</param>
+        /// <returns>If TResult is a boolean, the method returns <see langword="true"/> when the condition is true, and <see langword="false"/> otherwise.
+        /// If TResult is an object, the method returns the object when the condition evaluates to a value other than <see langword="null"/>.</returns>
+        /// <exception cref="ArgumentException">Thrown when TResult is not boolean or an object type.</exception>
+        Task<TResult> ExecuteAsync<TResult>(Func<Task<TResult>> condition);
     }
 }
